feat: #1 入力ホームの作成
